Remove wrong symlink check
Currently the code attempts to check whether the target of a symlink exists,
and does not extract the symlink if the target is deemed not to exist.

However, the check (`os.path.exists(os.path.join(target, i.linkname))`) is
buggy as it does not include the path of the symlink within target.

Even if this were fixed, the check would not be able to handle chains of
symlinks that happened to be iterated in the wrong order.

For now, remove the check entirely.